requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
Werkzeug==3.0.3
